CarIdentifier renaming file and moving it to the appropriate module
