+++ app/src/main/java/com/tomerpacific/caridentifier/screen/TirePressureScreen.kt
@@ -1,4 +1,4 @@
-package com.tomerpacific.caridentifier.composable
+package com.tomerpacific.caridentifier.screen
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
@@ -32,6 +32,7 @@ import androidx.compose.ui.text.style.TextDirection
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.tomerpacific.caridentifier.R
+import com.tomerpacific.caridentifier.composable.LoaderAnimation
 import com.tomerpacific.caridentifier.model.MainViewModel
 import com.tomerpacific.caridentifier.model.TirePressure
 import com.tomerpacific.caridentifier.model.allPropertiesNull
